move pydantic models to separate files
