+++ backend/models/Token.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class Token(BaseModel):
+    encoded_jwt: str
