+++ backend/models/User.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+
+class User(BaseModel):
+    username: str
+    name: str
+    timezone: int
+    bio: str
+    skills: list[str]
+    interests: list[str]
