+++ backend/models/__init__.py
@@ -1,0 +1,3 @@
+from . import Project
+from . import Token
+from . import User
